feat: move the fetch to component file
