+++ package.json
@@ -51,7 +51,6 @@
     "@testing-library/react": "^11.2.7",
     "@webpack-cli/serve": "^1.3.0",
     "autoprefixer": "^10.2.4",
-    "axios": "^1.7.9",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.1.0",
@@ -104,6 +103,7 @@
     "@headlessui/react": "^0.3.0",
     "@stripe/react-stripe-js": "^2.8.1",
     "@stripe/stripe-js": "^3.5.0",
+    "axios": "^1.7.9",
     "gapi-script": "^1.2.0",
     "i18next": "^19.1.0",
     "react-easy-crop": "^3.3.2",
